requirements.txt
torch
dm-control
hydra-core
tensorboard
matplotlib